update: rename useChatbotResponse hook
